requirements.txt
torch
torchvision
torchaudio
numpy
scipy
matplotlib
SoundFile
PyYAML
pystoi
pesq
h5py
python-sofa
pytest
